Improve handling of ValueOnly option in engine config
Enhanced the logic for managing the `ValueOnly` option in the engine configuration. The updated implementation ensures that the `ValueOnly` option is added, if missing, for current Lc0 binaries that supports this option. Additionally, it supports value head testing for Lc0 rewrite by passing in a valuehead argument in case newer binaries are in use.